refactor: consolidate interfaces
