+++ tests/indicators/_common/Observables/Stackoverflow.StaticSeries.Tests.cs
@@ -62,7 +62,7 @@ public class Stackoverflow : TestBase
 
         // act: clear provider cache (cascades to subscribers)
         int cutoff = qtyQuotes / 2;
-        provider.CascadeCacheRemoval(cutoff);
+        provider.RemoveRange(cutoff, true);
 
         provider.Quotes.Count.Should().Be(cutoff);
 
@@ -146,7 +146,7 @@ public class Stackoverflow : TestBase
 
         // act: clear provider cache (cascades to subscribers)
         int cutoff = qtyQuotes / 2;
-        provider.CascadeCacheRemoval(cutoff);
+        provider.RemoveRange(cutoff, true);
 
         provider.Quotes.Count.Should().Be(cutoff);
 
@@ -230,7 +230,7 @@ public class Stackoverflow : TestBase
 
         // act: clear provider cache (cascades to subscribers)
         int cutoff = qtyQuotes / 2;
-        provider.CascadeCacheRemoval(cutoff);
+        provider.RemoveRange(cutoff, true);
 
         provider.Quotes.Count.Should().Be(cutoff);
